Add JUnit and Mockito tests for Category and Product

+++ product/src/main/java/com/ims/product/entity/Product.java
@@ -28,12 +28,14 @@ public class Product {
 	private String sku;
 	
 	@Positive(message = "Price must be greater than zero")
-	private BigDecimal price;
+	private double price;
 	
 	@ManyToOne(cascade=CascadeType.ALL)
 	@JoinColumn(name = "category_id")
 	private Category categoryId;
 	
 	private Long supplierId;
+
+	
 
 }
